refactor(ui5-page): migrate to TS
